// src/payments/dto/webhook-payload.dto.ts
export interface PaystackWebhookPayload {
  event: string;
  data: {
    id: number;
    domain: string;
    status: string;
    reference: string;
    amount: number;
    message: string;
    gateway_response: string;
    paid_at: string;
    created_at: string;
    channel: string;
    currency: string;
    ip_address: string;
    metadata: {
      invoiceId: string;
      enrollmentId: string;
      custom_fields?: any[];
    };
    customer: {
      id: number;
      customer_code: string;
      email: string;
      first_name?: string;
      last_name?: string;
    };
    authorization?: {
      authorization_code: string;
      bin: string;
      last4: string;
      exp_month: string;
      exp_year: string;
      channel: string;
      card_type: string;
      bank: string;
      country_code: string;
      brand: string;
    };
  };
}